requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
firebase-admin==6.2.0
requests==2.31.0
python-dotenv==1.0.0 